--- enrich/summarize/metadata.json
@@ -1,84 +1,0 @@
-{
-  "timestamp": "2024-09-24T07-23-50-639Z",
-  "task": "Develop a text summarizer Flatfile Listener plugin:\n      - Create a RecordHook to summarize text fields\n      - Use the 'compromise' npm package for natural language processing\n      - Implement extractive summarization techniques\n      - Allow configuration of summary length or percentage\n      - Add summarized text to a new field in the record\n      - Give the user reasonable config options to specify the Sheet Slug, the Field(s) that are the text(s), whether the summarization should be done automatically",
-  "summary": "This code implements a Flatfile Listener plugin for text summarization. It uses the compromise library for natural language processing and provides configurable options for automatic summarization of content in specified fields.",
-  "steps": [
-    [
-      "Retrieve information about Flatfile Listeners and RecordHook.\n",
-      "#E1",
-      "PineconeAssistant",
-      "Provide information about Flatfile Listeners and RecordHook, including their structure and usage",
-      "Plan: Retrieve information about Flatfile Listeners and RecordHook.\n#E1 = PineconeAssistant[Provide information about Flatfile Listeners and RecordHook, including their structure and usage]"
-    ],
-    [
-      "Search for information about the 'compromise' npm package for natural language processing.\n",
-      "#E2",
-      "Google",
-      "compromise npm package natural language processing",
-      "Plan: Search for information about the 'compromise' npm package for natural language processing.\n#E2 = Google[compromise npm package natural language processing]"
-    ],
-    [
-      "Create the basic structure of the Flatfile Listener plugin with RecordHook.\n",
-      "#E3",
-      "LLM",
-      "Create a basic structure for a Flatfile Listener plugin with RecordHook, using the information from #E1",
-      "Plan: Create the basic structure of the Flatfile Listener plugin with RecordHook.\n#E3 = LLM[Create a basic structure for a Flatfile Listener plugin with RecordHook, using the information from #E1]"
-    ],
-    [
-      "Implement the text summarization logic using the 'compromise' package.\n",
-      "#E4",
-      "LLM",
-      "Implement extractive summarization techniques using the 'compromise' package based on the information from #E2 and integrate it into the Listener structure from #E3",
-      "Plan: Implement the text summarization logic using the 'compromise' package.\n#E4 = LLM[Implement extractive summarization techniques using the 'compromise' package based on the information from #E2 and integrate it into the Listener structure from #E3]"
-    ],
-    [
-      "Add configuration options for summary length or percentage and field selection.\n",
-      "#E5",
-      "LLM",
-      "Extend the Listener code from #E4 to include configuration options for summary length or percentage, Sheet Slug, and field selection",
-      "Plan: Add configuration options for summary length or percentage and field selection.\n#E5 = LLM[Extend the Listener code from #E4 to include configuration options for summary length or percentage, Sheet Slug, and field selection]"
-    ],
-    [
-      "Implement the logic to add the summarized text to a new field in the record.\n",
-      "#E6",
-      "LLM",
-      "Add code to the Listener from #E5 to create a new field in the record and populate it with the summarized text",
-      "Plan: Implement the logic to add the summarized text to a new field in the record.\n#E6 = LLM[Add code to the Listener from #E5 to create a new field in the record and populate it with the summarized text]"
-    ],
-    [
-      "Add configuration option for automatic summarization.\n",
-      "#E7",
-      "LLM",
-      "Extend the Listener code from #E6 to include a configuration option for automatic summarization",
-      "Plan: Add configuration option for automatic summarization.\n#E7 = LLM[Extend the Listener code from #E6 to include a configuration option for automatic summarization]"
-    ],
-    [
-      "Verify the Event Topics used in the Listener.\n",
-      "#E8",
-      "PineconeAssistant",
-      "Verify that the Event Topics used in the Listener are valid according to the event.topics.fact file",
-      "Plan: Verify the Event Topics used in the Listener.\n#E8 = PineconeAssistant[Verify that the Event Topics used in the Listener are valid according to the event.topics.fact file]"
-    ],
-    [
-      "Finalize the Flatfile Listener plugin code.\n",
-      "#E9",
-      "LLM",
-      "Combine all the components from #E3, #E4, #E5, #E6, #E7, and #E8 into a complete Flatfile Listener plugin code, ensuring all imports are used and the code is valid",
-      "Plan: Finalize the Flatfile Listener plugin code.\n#E9 = LLM[Combine all the components from #E3, #E4, #E5, #E6, #E7, and #E8 into a complete Flatfile Listener plugin code, ensuring all imports are used and the code is valid]"
-    ],
-    [
-      "Review and optimize the final code.\n",
-      "#E10",
-      "LLM",
-      "Review the code from #E9, remove any unused imports, validate plugin parameters, and ensure the code follows best practices for Flatfile Listener plugins",
-      "Plan: Review and optimize the final code.\n#E10 = LLM[Review the code from #E9, remove any unused imports, validate plugin parameters, and ensure the code follows best practices for Flatfile Listener plugins]"
-    ]
-  ],
-  "metrics": {
-    "tokens": {
-      "plan": 5077,
-      "state": 4675,
-      "total": 9752
-    }
-  }
-}
