--- import/faker/metadata.json
@@ -1,63 +1,0 @@
-{
-  "timestamp": "2024-09-24T08-56-21-753Z",
-  "task": "Create a Faker-generated example record Flatfile Listener plugin:\n      - Implement a custom action to generate thousands of example records using Faker\n      - Use the '@faker-js/faker' package to generate realistic fake data\n      - Analyze the sheet configuration to understand field types and constraints\n      - Generate contextually appropriate and diverse example data based on field types\n      - Allow configuration of the number of example records to generate (up to thousands)\n      - Implement efficient batch processing for generating large numbers of records\n      - Provide options for locale-specific data generation\n      - Include error handling for unsupported field types or constraints",
-  "summary": "This code implements a Flatfile Listener plugin that generates example records using Faker. It includes efficient batch processing, contextual data generation, error handling, and support for various field types. The plugin can be used to generate large numbers of realistic sample records based on the sheet configuration.",
-  "steps": [
-    [
-      "First, let's get information about Flatfile Listeners and the Record Hook plugin to understand the structure we'll be working with.\n",
-      "#E1",
-      "PineconeAssistant",
-      "Provide information about Flatfile Listeners and the Record Hook plugin structure",
-      "Plan: First, let's get information about Flatfile Listeners and the Record Hook plugin to understand the structure we'll be working with.\n#E1 = PineconeAssistant[Provide information about Flatfile Listeners and the Record Hook plugin structure]"
-    ],
-    [
-      "Next, we'll implement the custom action to generate example records using Faker.\n",
-      "#E3",
-      "LLM",
-      "Implement a custom action function that uses @faker-js/faker to generate example records based on the sheet configuration. Include parameters for the number of records to generate and locale options. Use the basic structure from #E2",
-      "Plan: Next, we'll implement the custom action to generate example records using Faker.\n#E3 = LLM[Implement a custom action function that uses @faker-js/faker to generate example records based on the sheet configuration. Include parameters for the number of records to generate and locale options. Use the basic structure from #E2]"
-    ],
-    [
-      "Now, let's add logic to analyze the sheet configuration and generate appropriate fake data based on field types.\n",
-      "#E4",
-      "LLM",
-      "Enhance the custom action function from #E3 to analyze sheet configuration and generate contextually appropriate fake data based on field types",
-      "Plan: Now, let's add logic to analyze the sheet configuration and generate appropriate fake data based on field types.\n#E4 = LLM[Enhance the custom action function from #E3 to analyze sheet configuration and generate contextually appropriate fake data based on field types]"
-    ],
-    [
-      "Implement efficient batch processing for generating large numbers of records.\n",
-      "#E5",
-      "LLM",
-      "Add batch processing logic to the custom action function from #E4 to efficiently generate thousands of records",
-      "Plan: Implement efficient batch processing for generating large numbers of records.\n#E5 = LLM[Add batch processing logic to the custom action function from #E4 to efficiently generate thousands of records]"
-    ],
-    [
-      "Add error handling for unsupported field types or constraints.\n",
-      "#E6",
-      "LLM",
-      "Implement error handling in the custom action function from #E5 for unsupported field types or constraints",
-      "Plan: Add error handling for unsupported field types or constraints.\n#E6 = LLM[Implement error handling in the custom action function from #E5 for unsupported field types or constraints]"
-    ],
-    [
-      "Finally, let's combine all the components and create the complete Faker-generated example record Flatfile Listener plugin.\n",
-      "#E7",
-      "LLM",
-      "Combine the components from #E2, #E3, #E4, #E5, and #E6 to create the complete Faker-generated example record Flatfile Listener plugin",
-      "Plan: Finally, let's combine all the components and create the complete Faker-generated example record Flatfile Listener plugin.\n#E7 = LLM[Combine the components from #E2, #E3, #E4, #E5, and #E6 to create the complete Faker-generated example record Flatfile Listener plugin]"
-    ],
-    [
-      "Validate the final code, check for unused imports, and ensure that the listener subscribes to valid Event Topics.\n",
-      "#E8",
-      "PineconeAssistant",
-      "Validate the Flatfile Listener plugin code from #E7, check for unused imports, and confirm that it subscribes to valid Event Topics",
-      "Plan: Validate the final code, check for unused imports, and ensure that the listener subscribes to valid Event Topics.\n#E8 = PineconeAssistant[Validate the Flatfile Listener plugin code from #E7, check for unused imports, and confirm that it subscribes to valid Event Topics]"
-    ]
-  ],
-  "metrics": {
-    "tokens": {
-      "plan": 5316,
-      "state": 5470,
-      "total": 10786
-    }
-  }
-}
