--- import/rss/metadata.json
@@ -1,84 +1,0 @@
-{
-  "timestamp": "2024-09-25T16-37-16-776Z",
-  "task": "Create a RSS Feed Import Flatfile Action:\n      - Implement a custom action to import new entries from an RSS feed into a Flatfile Sheet\n      - Allow users to specify the RSS feed URL and update frequency\n      - Parse the RSS feed and map feed items to appropriate sheet columns\n      - Implement deduplication to avoid importing duplicate entries\n      - Provide options for handling media enclosures and attachments\n      - Allow users to configure filters for importing specific types of content\n      - Implement error handling for feed parsing and network issues\n      - Generate a report of newly imported entries and any import failures\n      - Optionally, allow scheduling of regular imports",
-  "summary": "This code implements a Flatfile Action for importing RSS feed data, including error handling, scheduling, and reporting. It uses the Flatfile Listener to handle events, parses RSS feeds, maps data to sheet columns, generates reports, and schedules regular imports.",
-  "steps": [
-    [
-      "First, let's create a basic structure for our RSS Feed Import Action using Flatfile's Action framework.\n",
-      "#E1",
-      "PineconeAssistant",
-      "Example of a Flatfile Action structure for RSS Feed Import",
-      "Plan: First, let's create a basic structure for our RSS Feed Import Action using Flatfile's Action framework.\n#E1 = PineconeAssistant[Example of a Flatfile Action structure for RSS Feed Import]"
-    ],
-    [
-      "Now, let's implement the main logic for parsing the RSS feed and mapping items to sheet columns.\n",
-      "#E2",
-      "PineconeAssistant",
-      "Example code for parsing RSS feed and mapping to Flatfile sheet columns",
-      "Plan: Now, let's implement the main logic for parsing the RSS feed and mapping items to sheet columns.\n#E2 = PineconeAssistant[Example code for parsing RSS feed and mapping to Flatfile sheet columns]"
-    ],
-    [
-      "We need to implement deduplication to avoid importing duplicate entries. Let's use Flatfile's Record Hook for this.\n",
-      "#E3",
-      "PineconeAssistant",
-      "Example of using Record Hook in Flatfile for deduplication",
-      "Plan: We need to implement deduplication to avoid importing duplicate entries. Let's use Flatfile's Record Hook for this.\n#E3 = PineconeAssistant[Example of using Record Hook in Flatfile for deduplication]"
-    ],
-    [
-      "Let's add functionality to handle media enclosures and attachments from the RSS feed.\n",
-      "#E4",
-      "PineconeAssistant",
-      "Example of handling media enclosures and attachments in Flatfile",
-      "Plan: Let's add functionality to handle media enclosures and attachments from the RSS feed.\n#E4 = PineconeAssistant[Example of handling media enclosures and attachments in Flatfile]"
-    ],
-    [
-      "Implement user-configurable filters for importing specific types of content.\n",
-      "#E5",
-      "PineconeAssistant",
-      "Example of implementing content filters in Flatfile Action",
-      "Plan: Implement user-configurable filters for importing specific types of content.\n#E5 = PineconeAssistant[Example of implementing content filters in Flatfile Action]"
-    ],
-    [
-      "Add error handling for feed parsing and network issues.\n",
-      "#E6",
-      "PineconeAssistant",
-      "Example of error handling in Flatfile Action for RSS feed parsing and network issues",
-      "Plan: Add error handling for feed parsing and network issues.\n#E6 = PineconeAssistant[Example of error handling in Flatfile Action for RSS feed parsing and network issues]"
-    ],
-    [
-      "Generate a report of newly imported entries and any import failures.\n",
-      "#E7",
-      "PineconeAssistant",
-      "Example of generating import report in Flatfile Action",
-      "Plan: Generate a report of newly imported entries and any import failures.\n#E7 = PineconeAssistant[Example of generating import report in Flatfile Action]"
-    ],
-    [
-      "Implement scheduling functionality for regular imports.\n",
-      "#E8",
-      "PineconeAssistant",
-      "Example of implementing scheduled imports in Flatfile",
-      "Plan: Implement scheduling functionality for regular imports.\n#E8 = PineconeAssistant[Example of implementing scheduled imports in Flatfile]"
-    ],
-    [
-      "Combine all the components into a complete RSS Feed Import Action.\n",
-      "#E9",
-      "LLM",
-      "Combine the following components into a complete Flatfile Action for RSS Feed Import: #E1, #E2, #E3, #E4, #E5, #E6, #E7, #E8. Ensure all imports are correctly included and the code is valid.",
-      "Plan: Combine all the components into a complete RSS Feed Import Action.\n#E9 = LLM[Combine the following components into a complete Flatfile Action for RSS Feed Import: #E1, #E2, #E3, #E4, #E5, #E6, #E7, #E8. Ensure all imports are correctly included and the code is valid.]"
-    ],
-    [
-      "Validate the final Action code, checking for unused imports and ensuring all Event Topics are valid.\n",
-      "#E10",
-      "PineconeAssistant",
-      "Validate the following Flatfile Action code, remove unused imports, and confirm all Event Topics are valid: #E9",
-      "Plan: Validate the final Action code, checking for unused imports and ensuring all Event Topics are valid.\n#E10 = PineconeAssistant[Validate the following Flatfile Action code, remove unused imports, and confirm all Event Topics are valid: #E9]"
-    ]
-  ],
-  "metrics": {
-    "tokens": {
-      "plan": 4763,
-      "state": 5988,
-      "total": 10751
-    }
-  }
-}
